Refactored Animals to be less code
